+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/models/QueryThreadsRequest.kt
@@ -16,10 +16,45 @@
 
 package io.getstream.chat.android.client.api.models
 
+import io.getstream.chat.android.core.internal.InternalStreamChatApi
+import io.getstream.chat.android.models.FilterObject
+import io.getstream.chat.android.models.Thread
 import io.getstream.chat.android.models.User
+import io.getstream.chat.android.models.querysort.QuerySortByField
+import io.getstream.chat.android.models.querysort.QuerySortByField.Companion.descByName
+import io.getstream.chat.android.models.querysort.QuerySorter
 
 /**
  * Query threads request.
+ *
+ * @property filter The filter object for the query. Supported fields:
+ *  - `channel_cid` Filter by channel CID. Supported operators: `$eq`, `$in`
+ *  - `channel.disabled` Filter by channel disabled status. Supported operators: `$eq`
+ *  - `parent_message_id` Filter by parent message ID. Supported operators: `$eq`, `$in`
+ *  - `created_by_user_id` Filter by thread creator’s user ID. Supported operators: `$eq`, `$in`
+ *  - `created_at` Filter by thread creation timestamp. Supported operators: `$eq`, `$gt`, `$lt`, `$gte`, `$lte`
+ *  - `updated_at` Filter by thread update timestamp. Supported operators: `$eq`, `$gt`, `$lt`, `$gte`, `$lte`
+ *  - `last_message_at` Filter by last message timestamp. Supported operators: `$eq`, `$gt`, `$lt`, `$gte`, `$lte`
+ *
+ * For more info,
+ * see [Filtering and Sorting Threads](https://getstream.io/chat/docs/android/threads/#filtering-and-sorting-threads).
+ *
+ * @property sort The sort object for the query. Supported fields:
+ *  - `active_participant_count` Number of active participants in the thread
+ *  - `created_at` Thread creation timestamp
+ *  - `last_message_at` Timestamp of the last message in the thread
+ *  - `parent_message_id` ID of the parent message
+ *  - `participant_count` Total number of participants in the thread
+ *  - `reply_count` Number of replies in the thread
+ *  - `updated_at` Thread last update timestamp
+ *
+ *  If not provided, threads will be sorted by:
+ *  1. Unread status (unread threads first)
+ *  2. Last message timestamp (newest first)
+ *  3. Parent message ID (descending)
+ *
+ * For more info,
+ * see [Filtering and Sorting Threads](https://getstream.io/chat/docs/android/threads/#filtering-and-sorting-threads).
  *
  * @property watch If true, all the channels corresponding to threads returned in response will be watched.
  * Defaults to true.
@@ -29,10 +64,14 @@ import io.getstream.chat.android.models.User
  * @property participantLimit The number of thread participants to request per thread. Defaults to 100.
  * @property prev The previous pagination token. This token can be used to fetch the previous page of threads.
  * @property replyLimit The number of latest replies to fetch per thread. Defaults to 2.
- * @property user The user for which the threads are queried. Defaults to null.
- * @property userId The user ID for which the threads are queried. Defaults to null.
+ * @property user The user for which the threads are queried. Defaults to null. This property is deprecated and is not
+ * used.
+ * @property userId The user ID for which the threads are queried. Defaults to null. This property is deprecated and is
+ * not used.
  */
 public data class QueryThreadsRequest @JvmOverloads constructor(
+    public val filter: FilterObject? = null,
+    public val sort: QuerySorter<Thread> = DefaultSort,
     public val watch: Boolean = true,
     public val limit: Int = 10,
     public val memberLimit: Int = 100,
@@ -40,6 +79,24 @@ public data class QueryThreadsRequest @JvmOverloads constructor(
     public val participantLimit: Int = 100,
     public val prev: String? = null,
     public val replyLimit: Int = 2,
+    @Deprecated(message = "Do not use this property, it is ignored in the query.")
     public val user: User? = null,
+    @Deprecated(message = "Do not use this property, it is ignored in the query.")
     public val userId: String? = null,
-)
+) {
+
+    public companion object {
+
+        /**
+         * The default sort order for threads:
+         * 1. Unread status (unread threads first)
+         * 2. Last message timestamp (newest first)
+         * 3. Parent message ID (descending)
+         */
+        @InternalStreamChatApi
+        public val DefaultSort: QuerySorter<Thread> = QuerySortByField
+            .descByName<Thread>("has_unread")
+            .descByName("last_message_at")
+            .descByName("parent_message_id")
+    }
+}